Minimal changes to enable building and running
README.md has more details on the required packages and instructions
to build and run.

32-bit PulseAudio, X11, X11 extensions and the 32-bit runtime are
needed.
The game requires an 8-bit PseudoColor compatible display with
a resolution of 320x200, which can be gotten with Xephyr.
The game uses OSS audio, which can be used with the PulseAudio OSS
Wrapper.

Mouse controls don't seem to work.
Music doesn't work (functions are stubs).
Network not tested, presumed not to work.
Screen size and resolution cannot be changed.